fetching the data both server and client side works as expected, however there is a problem with rehydrating from the sdk. When we stringify the SDK to make sure we can transmit it to the client we lose the functionality in there and receive a plain JSON object. Then when we try to read the object the same way we do with the complete SDK response object we get errors as the data is not in the correct location and the functions are no longer present.
Perhaps in order to resolve this we could either:
a) not use the SDK and go straight to hapi.
b) add a rehydrating method to the SDK which allows us to pipe this data back in and recieve the complete objects back.

(a) would be simpler, (b) would be more complete.